[DESIGN/#26] Refactor: GenState 함수의 위치 변경 및 쿠키 이름 통일
